+++ cmd/root.go
@@ -23,13 +23,12 @@ var (
 
 // rootCmd represents the base command when called without any subcommands
 var (
-	rootCmd = &cobra.Command{
+	greetings = fmt.Sprintf("\nHossted CLI %s - for help please contact us at support@hossted.com\n\n", VERSION)
+	rootCmd   = &cobra.Command{
 		Use:     "hossted",
 		Version: VERSION,
-		Short:   "A brief description of your application.",
-		Long: `
-A brief description of your application
-`,
+		Short:   greetings,
+		Long:    greetings,
 	}
 )
 
@@ -108,7 +107,6 @@ func checkConfigFilePath() (string, error) {
 
 		fmt.Printf("\nNo existing config file. \nNew config file is created  - %s \n\n", cfgPath)
 
-		return "", err
 	} else {
 		// Normal case
 		// Do nothing. config.yaml exists